requirements.txt
streamlit>=1.32.0
plotly>=5.18.0
pandas>=2.0.0
python-nmap>=0.7.1
